Rename widget configurations that options on create
