+++ graylog2-web-interface/src/components/widgets/configurations/CountWidgetEditConfiguration.jsx
@@ -3,7 +3,7 @@ import { Input } from 'react-bootstrap';
 
 import { QueryConfiguration } from 'components/widgets/configurations';
 
-const CountWidgetConfiguration = React.createClass({
+const CountWidgetEditConfiguration = React.createClass({
   propTypes: {
     config: PropTypes.object.isRequired,
     onChange: PropTypes.func.isRequired,
@@ -41,4 +41,4 @@ const CountWidgetConfiguration = React.createClass({
   },
 });
 
-export default CountWidgetConfiguration;
+export default CountWidgetEditConfiguration;
