+++ graylog2-web-interface/src/components/widgets/configurations/index.js
@@ -1,6 +1,6 @@
-export { default as CountWidgetConfiguration } from './CountWidgetConfiguration';
+export { default as CountWidgetEditConfiguration } from './CountWidgetEditConfiguration';
 export { default as FieldChartWidgetConfiguration } from './FieldChartWidgetConfiguration';
 export { default as QueryConfiguration } from './QueryConfiguration';
-export { default as QuickValuesWidgetConfiguration } from './QuickValuesWidgetConfiguration';
+export { default as QuickValuesWidgetEditConfiguration } from './QuickValuesWidgetEditConfiguration';
 export { default as StackedChartWidgetConfiguration } from './StackedChartWidgetConfiguration';
-export { default as StatisticalCountWidgetConfiguration } from './StatisticalCountWidgetConfiguration';
+export { default as StatisticalCountWidgetEditConfiguration } from './StatisticalCountWidgetEditConfiguration';
